sales_research_app/requirements.txt
streamlit
google-auth
google-auth-oauthlib
google-api-python-client
google-generativeai
trafilatura
markdownify
lxml_html_clean